+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### next
+- fix `:trash` internal not working on staged files
+
 ### v1.30.2 - 2023-12-23
 <a name="v1.30.2"></a>
 - don't canonicalize paths on windows on new panels - Fix #809